Added twitter libraries
I've updated the CSS a bit and added more content on pages. I've also added an embedded twitter search and started implementing the twitter REST api using the linqtotwitter library, the idea is to let people post to twitter from the website if they see something interesting in the news.